requirements.txt
django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1